update jspm instructions
This updates the jspm installation instructions to remove `npm:` registry prefix as it is no longer necessary as the package is now mapped in the jspm registry.

Additionally. the example `tsconfig.json` now reflects the path corresponding to the latest package version and uses the lowercase module kind name `"system"` to match the other example and to avoid warnings issued by VS Code when `"System"` is specified.